+++ app/src/main/java/com/example/studyapp/ui/components/study/Dialogs.kt
@@ -34,145 +34,5 @@ import com.example.studyapp.R
 import com.example.studyapp.data.Subtopic
 import com.example.studyapp.ui.components.FullScreenDialog
 
-@Composable
-fun SubtopicFullScreenDialog(
-    titleRes: Int,
-    onDismiss: () -> Unit,
-    saveSubtopic: (String, String, String?) -> Unit,
-    modifier: Modifier = Modifier,
-    subtopic: Subtopic? = null,
-) {
-    var title by rememberSaveable { mutableStateOf(subtopic?.title ?: "") }
-    var description by rememberSaveable { mutableStateOf(subtopic?.description ?: "") }
-    var imageUri by rememberSaveable { mutableStateOf(subtopic?.imageUri ?: "") }
-    FullScreenDialog(
-        titleRes = titleRes,
-        onDismiss = onDismiss,
-        onConfirm = { saveSubtopic(title, description, imageUri) },
-        modifier = modifier.padding(horizontal = 16.dp)
-    ) { innerPadding ->
-        SubtopicInputFields(
-            updateTitle = { title = it },
-            updateDescription = { description = it },
-            updateImageUri = { imageUri = it },
-            title = title,
-            description = description,
-            imageUri = imageUri,
-            modifier = Modifier
-                .padding(paddingValues = innerPadding)
-                .padding(horizontal = 8.dp)
-                .fillMaxWidth(),
-        )
-    }
-}
 
-@Composable
-fun SubtopicDialog(
-    titleRes: Int,
-    onDismiss: () -> Unit,
-    saveSubtopic: (String, String, String?) -> Unit,
-    modifier: Modifier = Modifier,
-    subtopic: Subtopic? = null,
-) {
-    var title by rememberSaveable { mutableStateOf(subtopic?.title ?: "") }
-    var description by rememberSaveable { mutableStateOf(subtopic?.description ?: "") }
-    var imageUri by rememberSaveable { mutableStateOf(subtopic?.imageUri ?: "") }
-    AlertDialog(
-        // Dialog should not close when clicking outside.
-        onDismissRequest = {},
-        title = { Text(stringResource(titleRes)) },
-        text = {
-            SubtopicInputFields(
-                updateTitle = { title = it },
-                updateDescription = { description = it },
-                updateImageUri = { imageUri = it },
-                title = title,
-                description = description,
-                imageUri = imageUri,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 8.dp),
-            )
-        },
-        confirmButton = {
-            TextButton(onClick = {
-                saveSubtopic(title, description, imageUri)
-                onDismiss()
-            }
-            ) {
-                Text(stringResource(R.string.save))
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = { onDismiss() }) {
-                Text(stringResource(R.string.cancel))
-            }
-        },
-        modifier = modifier
-    )
-}
-
-@Composable
-private fun SubtopicInputFields(
-    updateTitle: (String) -> Unit,
-    updateDescription: (String) -> Unit,
-    updateImageUri: (String) -> Unit,
-    title: String,
-    description: String,
-    imageUri: String,
-    modifier: Modifier = Modifier,
-) {
-    val focusManager = LocalFocusManager.current
-    val keyboardController = LocalSoftwareKeyboardController.current
-    val scrollState = rememberScrollState()
-    if (scrollState.isScrollInProgress) {
-        keyboardController?.hide()
-    }
-    val pickMedia = rememberLauncherForActivityResult(PickVisualMedia()) { uri ->
-        updateImageUri(uri.toString())
-    }
-    Column(
-        modifier = modifier
-            .verticalScroll(state = scrollState)
-            .pointerInput(Unit) {
-                detectTapGestures(onTap = { focusManager.clearFocus() })
-            },
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        OutlinedTextField(
-            value = title,
-            onValueChange = { updateTitle(it) },
-            modifier = Modifier.fillMaxWidth(),
-            label = { Text(stringResource(R.string.title)) },
-        )
-        OutlinedTextField(
-            value = description,
-            onValueChange = { updateDescription(it) },
-            modifier = Modifier.fillMaxWidth(),
-            label = { Text(stringResource(R.string.description)) }
-        )
-        OutlinedTextField(
-            value = imageUri,
-            onValueChange = { updateImageUri(it) },
-            modifier = Modifier.fillMaxWidth(),
-            trailingIcon = {
-                Icon(
-                    painter = painterResource(R.drawable.outline_image_24),
-                    contentDescription = stringResource(R.string.image),
-                    modifier = Modifier
-                        .size(24.dp)
-                        .clickable {
-                            pickMedia.launch(PickVisualMediaRequest(PickVisualMedia.ImageOnly))
-                        }
-                )
-            },
-            label = { Text(stringResource(R.string.image)) }
-        )
-        AsyncImage(
-            model = imageUri,
-            contentDescription = null,
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
