+++ app/src/main/java/com/example/studyapp/ui/screens/subtopics/SubtopicsScreen.kt
@@ -1,0 +1,187 @@
+package com.example.studyapp.ui.screens.subtopics
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.PreviewLightDark
+import androidx.compose.ui.tooling.preview.PreviewScreenSizes
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.window.core.layout.WindowSizeClass.Companion.WIDTH_DP_MEDIUM_LOWER_BOUND
+import com.example.studyapp.R
+import com.example.studyapp.data.Subtopic
+import com.example.studyapp.data.Topic
+import com.example.studyapp.ui.theme.StudyAppTheme
+import com.example.studyapp.ui.viewmodels.SubtopicsViewModel
+
+@Composable
+fun SubtopicsScreen(
+    subtopicsViewModel: SubtopicsViewModel,
+    navigateToSubtopic: (Int) -> Unit,
+    navigateToTopic: (Int) -> Unit,
+    navigateBack: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val topic by subtopicsViewModel.topic.collectAsStateWithLifecycle()
+    val subtopics by subtopicsViewModel.subtopics.collectAsStateWithLifecycle()
+    val topics by subtopicsViewModel.topics.collectAsStateWithLifecycle()
+    SubtopicsScreen(
+        subtopics = subtopics,
+        topics = topics,
+        saveSubtopic = { title, description, imageUri ->
+            subtopicsViewModel.createSubtopic(
+                title = title, description = description, imageUri = imageUri
+            )
+        },
+        deleteTopic = subtopicsViewModel::deleteTopic,
+        updateTopic = subtopicsViewModel::updateTopic,
+        navigateToSubtopic = navigateToSubtopic,
+        navigateToTopic = navigateToTopic,
+        navigateBack = navigateBack,
+        topic = topic ?: return,
+        modifier = modifier
+    )
+}
+
+@Composable
+private fun SubtopicsScreen(
+    subtopics: List<Subtopic>?,
+    topics: List<Topic>?,
+    topic: Topic?,
+    saveSubtopic: (String, String, String?) -> Unit,
+    deleteTopic: () -> Unit,
+    updateTopic: (Topic) -> Unit,
+    navigateToSubtopic: (Int) -> Unit,
+    navigateToTopic: (Int) -> Unit,
+    navigateBack: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    var showFullScreenDialog by rememberSaveable { mutableStateOf(false) }
+    var showBasicDialog by rememberSaveable { mutableStateOf(false) }
+    val windowSizeClass = currentWindowAdaptiveInfo().windowSizeClass
+
+    if (topic == null || subtopics == null || topics == null) {
+        Box(modifier = modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+            CircularProgressIndicator()
+        }
+    } else {
+        if (showFullScreenDialog) {
+            SubtopicFullScreenDialog(
+                titleRes = R.string.create_subtopic,
+                onDismiss = { showFullScreenDialog = false },
+                saveSubtopic = { title, description, imageUri ->
+                    saveSubtopic(title, description, imageUri)
+                    showFullScreenDialog = false
+                },
+                modifier = modifier
+            )
+        } else {
+            if (showBasicDialog) {
+                SubtopicDialog(
+                    titleRes = R.string.create_subtopic,
+                    onDismiss = { showBasicDialog = false },
+                    saveSubtopic = { title, description, imageUri ->
+                        saveSubtopic(title, description, imageUri)
+                        showBasicDialog = false
+                    },
+                    modifier = modifier
+                )
+            }
+            SubtopicsScaffold(
+                subtopics = subtopics,
+                topics = topics,
+                topic = topic,
+                onCreateSubtopic = {
+                    // Basic dialog for medium and expanded screen width
+                    if (windowSizeClass.isWidthAtLeastBreakpoint(WIDTH_DP_MEDIUM_LOWER_BOUND)) {
+                        showBasicDialog = true
+                    }
+                    // Fullscreen dialog for compact screen width
+                    else {
+                        showFullScreenDialog = true
+                    }
+                },
+                deleteTopic = deleteTopic,
+                updateTopic = updateTopic,
+                navigateToSubtopic = navigateToSubtopic,
+                navigateToTopic = navigateToTopic,
+                navigateBack = navigateBack,
+                modifier = modifier.fillMaxWidth()
+            )
+        }
+    }
+}
+
+@PreviewLightDark
+@PreviewScreenSizes
+@Composable
+private fun SubtopicsScreenPreview() {
+    StudyAppTheme {
+        SubtopicsScreen(
+            subtopics = listOf(
+                Subtopic(
+                    id = 1,
+                    topicId = 0,
+                    title = "Subtopic 1",
+                    description = "Description 1",
+                    checked = false,
+                    bookmarked = false,
+                    imageUri = null
+                ), Subtopic(
+                    id = 2,
+                    topicId = 0,
+                    title = "Subtopic 2",
+                    description = "Description 2",
+                    checked = true,
+                    bookmarked = true,
+                    imageUri = null
+                ), Subtopic(
+                    id = 3,
+                    topicId = 0,
+                    title = "Subtopic 3",
+                    description = "Description 3",
+                    checked = false,
+                    bookmarked = false,
+                    imageUri = null
+                )
+            ),
+            topics = listOf(Topic(id = 1, title = "Topic 1", checked = false)),
+            topic = Topic(
+                id = 1, title = "Android Taint Analysis", checked = false
+            ),
+            navigateToSubtopic = {},
+            navigateToTopic = {},
+            navigateBack = {},
+            saveSubtopic = { _, _, _ -> },
+            deleteTopic = {},
+            updateTopic = {})
+    }
+}
+
+@PreviewLightDark
+@PreviewScreenSizes
+@Composable
+private fun LoadingScreenPreview() {
+    StudyAppTheme {
+        SubtopicsScreen(
+            subtopics = null,
+            topics = null,
+            topic = null,
+            navigateToSubtopic = {},
+            navigateToTopic = {},
+            navigateBack = {},
+            saveSubtopic = { _, _, _ -> },
+            deleteTopic = {},
+            updateTopic = {},
+        )
+    }
+}
+
